+++ JavaScript/index.js
@@ -1,4 +1,4 @@
-//1번 페이지 이동
+//페이지 이동
 const homeBtn = document.querySelector(".home-btn");
 
 homeBtn.addEventListener("click", function () {
@@ -7,79 +7,60 @@ homeBtn.addEventListener("click", function () {
   window.location.href = homePageUrl;
 });
 
-let total = 0;
-//5번 로컬스토리지 정보 가져오기
+// //상수 데이터 랜더링하기
+import { itemList } from "./constants.js";
+
+// //상수 데이터 > 전체 상품 넣기
+const productContainer = function () {
+  const productSection = document.querySelector(".wishList");
+  const productImg = itemList.map((item) => {
+    return `
+    <article class = "wishItem" data-category="${item.category}" data-id="${
+      item.id
+    }">
+      <img src=${`img/${item.category}/${item.img}`}>
+      <span class="likeBtn">♥</span>
+      <h4>${item.item_name}</h4>
+      <p>${item.item_price}</p>
+    </article>
+    `;
+  });
+  productSection.innerHTML = productImg.join("");
+};
+productContainer();
+
+// item 클릭 시 알럿 노출
 document.addEventListener("DOMContentLoaded", function () {
-  const tbody = document.querySelector("#cart-table tbody");
+  const SHOPPING_LIST = document.querySelectorAll(".wishItem");
+  SHOPPING_LIST.forEach(function (productList) {
+    productList.addEventListener("click", function (event) {
+      const itemId = productList.getAttribute("data-id");
+      const imageSrc = productList.querySelector("img").src;
+      const itemName = productList.querySelector("h4").textContent;
+      const itemPrice = productList.querySelector("p").textContent;
+      const itemCategory = productList.getAttribute("data-category");
 
-  function addListToCart() {
-    const confirmedItems = JSON.parse(localStorage.getItem("item")); // []
-    confirmedItems.forEach((confirmedItem) => {
-      total += Number(confirmedItem.price);
-      const addNewRow = document.createElement("tr");
+      //확인버튼 눌렀을 경우
+      const confirmed = confirm(`장바구니에 ${itemName} 담기!`);
 
-      const addItemCheckbox = document.createElement("td");
-      const checkbox = document.createElement("input");
-      checkbox.type = "checkbox";
-      checkbox.name = "item-checkbox";
-      addItemCheckbox.appendChild(checkbox);
-
-      const addItemImg = document.createElement("td");
-      const itemImgSrc = document.createElement("img");
-      itemImgSrc.src = confirmedItem.image;
-      addItemImg.appendChild(itemImgSrc);
-
-      const addItemName = document.createElement("td");
-      addItemName.textContent = confirmedItem.name;
-
-      const addItemPrice = document.createElement("td");
-      addItemPrice.textContent = confirmedItem.price;
-
-      const addItemCategory = document.createElement("td");
-      addItemCategory.textContent = confirmedItem.categoryName;
-
-      const addItemBuyBtn = document.createElement("td");
-      const BuyBtn = document.createElement("input");
-      BuyBtn.type = "button";
-      BackBtn.name = "item-buy";
-      BuyBtn.value = "삭제";
-      BuyBtn.addEventListener("click", () => {
-        const currentId = confirmedItem.id;
+      if (confirmed) {
+        const productListInfo = {
+          id: itemId,
+          name: itemName,
+          price: itemPrice,
+          image: imageSrc,
+          categoryName: itemCategory,
+        };
 
         const cartItem = JSON.parse(localStorage.getItem("item")) || []; // [] 기존배열
-        cartItem.forEach((item, index) => {
-          console.log("idx", index, "item", item);
-          if (item.id === currentId) {
-            cartItem.splice(index, 1);
-          }
-        });
+        cartItem.push(productListInfo); // 기존배열 + 새아이템 => []
         localStorage.setItem("item", JSON.stringify(cartItem));
-        location.reload();
-      });
-      addItemBuyBtn.appendChild(BuyBtn);
+        alert("장바구니 담기 완료!");
 
-      addNewRow.append(
-        addItemCheckbox,
-        addItemImg,
-        addItemName,
-        addItemPrice,
-        addItemCategory,
-        addItemBuyBtn
-      );
-
-      tbody.appendChild(addNewRow);
-      console.log(confirmedItem);
+        // window.location.href = "cart.html";
+      } else {
+        alert("장바구니 담기가 취소되었습니다");
+      }
     });
-    console.log(total);
-  }
-  console.log(tbody);
-  addListToCart();
-});
-
-const cartToHomeBtn = document.querySelector(".go-to-home-button");
-
-cartToHomeBtn.addEventListener("click", function () {
-  const TohomePageUrl = "index.html";
-
-  window.location.href = TohomePageUrl;
+  });
 });
