+++ cart.html
@@ -9,6 +9,10 @@
       rel="stylesheet"
       href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@20..48,100..700,0..1,-50..200"
     />
+    <link
+      rel="stylesheet"
+      href="https://fonts.googleapis.com/css2?family=Material+Symbols+Outlined:opsz,wght,FILL,GRAD@24,400,0,0"
+    />
   </head>
   <body>
     <aside class="sidebar">
@@ -45,11 +49,24 @@
         <button class="buy-button">넌 내거야</button>
         <button class="go-to-home-button">홈으로</button>
       </article>
+      <!--모달 영역-->
+      <div class="buyModal">
+        <div class="buyModalPopup">
+          <span class="material-symbols-outlined"> close </span>
+          <h3>지민이에게 선물하기</h3>
+          <div class="modalBuyListArea">
+            <span class="buyListArea"></span>
+          </div>
+          <button type="button" class="buyModalButton">구매하기</button>
+        </div>
+      </div>
+      <!--모달 영역-->
     </section>
-    <script src="./JavaScript/index.js"></script>
-    <script src="./JavaScript/constants.js"></script>
-    <script src="./JavaScript/sidebar.js"></script>
-    <script src="./JavaScript/wishlist.js"></script>
-    <script src="./JavaScript/nav.js"></script>
+    <script src="./JavaScript/index.js" type="module"></script>
+    <script src="./JavaScript/constants.js" type="module"></script>
+    <script src="./JavaScript/sidebar.js" type="module"></script>
+    <script src="./JavaScript/wishlist.js" type="module"></script>
+    <script src="./JavaScript/nav.js" type="module"></script>
+    <script src="./JavaScript/modal.js" type="module"></script>
   </body>
 </html>
