+++ index.html
@@ -127,10 +127,11 @@
       <li>최고의 1주차 금잔디조 : 최주용 유태승 문성희</li>
       <li>불철주야 함께 해준 파짱님: 이재훈</li>
     </footer>
-    <script src="./JavaScript/index.js"></script>
+    <script src="./JavaScript/index.js" type="module"></script>
     <script src="./JavaScript/constants.js" type="module"></script>
-    <script src="./JavaScript/sidebar.js"></script>
-    <script src="./JavaScript/nav.js"></script>
+    <script src="./JavaScript/sidebar.js" type="module"></script>
+    <script src="./JavaScript/nav.js" type="module"></script>
     <script src="./JavaScript/wishlist.js" type="module"></script>
+    <script src="./JavaScript/modal.js" type="module"></script>
   </body>
 </html>
